+++ lib/plausible/session/write_buffer.ex
@@ -1,83 +1,36 @@
 defmodule Plausible.Session.WriteBuffer do
-  use GenServer
-  require Logger
-
-  alias Plausible.IngestRepo
-
-  def start_link(_opts) do
-    GenServer.start_link(__MODULE__, [], name: __MODULE__)
+  def child_spec(opts) do
+    opts = Keyword.merge([name: __MODULE__, schema: Plausible.ClickhouseSessionV2], opts)
+    Plausible.Ingestion.WriteBuffer.child_spec(opts)
   end
 
-  def init(buffer) do
-    Process.flag(:trap_exit, true)
-    timer = Process.send_after(self(), :tick, flush_interval_ms())
-    {:ok, %{buffer: buffer, timer: timer}}
+  def start_link(opts) do
+    opts = Keyword.merge([name: __MODULE__, schema: Plausible.ClickhouseSessionV2], opts)
+    Plausible.Ingestion.WriteBuffer.start_link(opts)
   end
 
+  @spec insert(sessions) :: sessions when sessions: [%Plausible.ClickhouseSessionV2{}]
   def insert(sessions) do
-    GenServer.cast(__MODULE__, {:insert, sessions})
-    {:ok, sessions}
+    sessions =
+      Enum.map(sessions, fn %{is_bounce: is_bounce} = session ->
+        is_bounce =
+          case is_bounce do
+            true -> 1
+            false -> 0
+            other -> other
+          end
+
+        %{session | is_bounce: is_bounce}
+      end)
+
+    Plausible.Ingestion.WriteBuffer.insert(
+      __MODULE__,
+      Plausible.ClickhouseSessionV2,
+      sessions
+    )
   end
 
-  def flush() do
-    GenServer.call(__MODULE__, :flush, :infinity)
-    :ok
-  end
-
-  def handle_cast({:insert, sessions}, %{buffer: buffer} = state) do
-    new_buffer = sessions ++ buffer
-
-    if length(new_buffer) >= max_buffer_size() do
-      Logger.info("Buffer full, flushing to disk")
-      Process.cancel_timer(state[:timer])
-      do_flush(new_buffer)
-      new_timer = Process.send_after(self(), :tick, flush_interval_ms())
-      {:noreply, %{buffer: [], timer: new_timer}}
-    else
-      {:noreply, %{state | buffer: new_buffer}}
-    end
-  end
-
-  def handle_info(:tick, %{buffer: buffer}) do
-    do_flush(buffer)
-    timer = Process.send_after(self(), :tick, flush_interval_ms())
-    {:noreply, %{buffer: [], timer: timer}}
-  end
-
-  def handle_call(:flush, _from, %{buffer: buffer} = state) do
-    Process.cancel_timer(state[:timer])
-    do_flush(buffer)
-    new_timer = Process.send_after(self(), :tick, flush_interval_ms())
-    {:reply, nil, %{buffer: [], timer: new_timer}}
-  end
-
-  def terminate(_reason, %{buffer: buffer}) do
-    Logger.info("Flushing session buffer before shutdown...")
-    do_flush(buffer)
-  end
-
-  defp do_flush(buffer) do
-    case buffer do
-      [] ->
-        nil
-
-      sessions ->
-        Logger.info("Flushing #{length(sessions)} sessions")
-
-        sessions =
-          sessions
-          |> Enum.map(&(Map.from_struct(&1) |> Map.delete(:__meta__)))
-          |> Enum.reverse()
-
-        IngestRepo.insert_all(Plausible.ClickhouseSessionV2, sessions)
-    end
-  end
-
-  defp flush_interval_ms() do
-    Keyword.fetch!(Application.get_env(:plausible, IngestRepo), :flush_interval_ms)
-  end
-
-  defp max_buffer_size() do
-    Keyword.fetch!(Application.get_env(:plausible, IngestRepo), :max_buffer_size)
+  def flush do
+    Plausible.Ingestion.WriteBuffer.flush(__MODULE__)
   end
 end
